+++ clients/admin-ui/src/features/datastore-connections/datastore-connection.slice.ts
@@ -4,6 +4,7 @@ import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 import type { RootState } from "../../app/store";
 import { BASE_URL, CONNECTION_ROUTE } from "../../constants";
 import { selectToken } from "../auth";
+import { addCommonHeaders } from "../common/CommonHeaders";
 import {
   ConnectionType,
   DatastoreConnection,
@@ -140,11 +141,8 @@ export const datastoreConnectionApi = createApi({
   baseQuery: fetchBaseQuery({
     baseUrl: BASE_URL,
     prepareHeaders: (headers, { getState }) => {
-      const token = selectToken(getState() as RootState);
-      headers.set("Access-Control-Allow-Origin", "*");
-      if (token) {
-        headers.set("authorization", `Bearer ${token}`);
-      }
+      const token: string | null = selectToken(getState() as RootState);
+      addCommonHeaders(headers, token);
       return headers;
     },
   }),